+++ src/Molecules/Input/FormField/FormField.stories.tsx
@@ -1,29 +1,47 @@
 import { storiesOf } from '@storybook/react';
 import * as React from 'react';
+import { Flexbox } from '../../..';
 
 // importing from this folder
 // since we are not exporting this component itself
 import { FormField } from '.';
 
 storiesOf('Molecules | Input / FormField', module)
-  .add('Docs', () => <p>This is a helper component, you are probably looking for something else</p>)
+  .add('Docs', () => (
+    <p>
+      This component provides a label including the required look of the label. It also provides the
+      displaying of error message and additional information.
+    </p>
+  ))
   .add('Default', () => (
-    <FormField label="Label">
-      <div style={{ background: 'aqua', width: '100%' }}>Pass in any children you want</div>
+    <FormField label="Label" fieldId="unique-id-1">
+      <div style={{ background: 'aqua' }}>Pass in any children you want</div>
+    </FormField>
+  ))
+  .add('With group of items', () => (
+    <FormField label="Caption to a group of items" fieldId="unique-id-1" group>
+      <Flexbox container gutter={5}>
+        <Flexbox item>
+          <div style={{ background: 'aqua' }}>Pass in any children you want</div>
+        </Flexbox>
+        <Flexbox item>
+          <div style={{ background: 'aqua' }}>Pass in any children you want</div>
+        </Flexbox>
+      </Flexbox>
     </FormField>
   ))
   .add('Required', () => (
-    <FormField label="Label" required>
+    <FormField label="Label" fieldId="unique-id-1" required>
       <div style={{ background: 'aqua' }}>Pass in any children you want</div>
     </FormField>
   ))
   .add('Error', () => (
-    <FormField label="Label" error="error">
-      <div style={{ background: 'aqua', width: '100%' }}>Pass in any children you want</div>
+    <FormField label="Label" error="error" fieldId="unique-id-2">
+      <div style={{ background: 'aqua' }}>Pass in any children you want</div>
     </FormField>
   ))
   .add('Extra info', () => (
-    <FormField label="Label" extraInfo="Some extra info">
-      <div style={{ background: 'aqua', width: '100%' }}>Pass in any children you want</div>
+    <FormField label="Label" extraInfo="Some extra info" fieldId="unique-id-2">
+      <div style={{ background: 'aqua' }}>Pass in any children you want</div>
     </FormField>
   ));