+++ src/Molecules/Input/FormField/FormField.tsx
@@ -1,143 +1,74 @@
 import React from 'react';
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 import { motion, AnimatePresence } from 'framer-motion';
-import R from 'ramda';
 import { Props } from './FormField.types';
-
-import { VisuallyHidden, Flexbox, Typography } from '../../..';
-import { visuallyHiddenCss as visuallyHidden } from '../../../Atoms/VisuallyHidden';
+import { VisuallyHidden, FormLabel, Typography, Fieldset, Legend } from '../../..';
+import { assert } from '../../../common/utils';
 
 const hasError = (error?: Props['error']) => error && error !== '';
-
-const LINE_HEIGHT_INFO_BELOW = 17;
-
-const height = css<Pick<Props, 'size'>>`
-  height: ${p => (p.size === 's' ? p.theme.spacing.unit(8) : p.theme.spacing.unit(10))}px;
-`;
 
 const Wrapper = styled.div<{ width?: string | number }>`
   ${p => (p.width ? `width: ${p.width};` : 'width: 200px;')}
   display: inline-block;
 `;
 
-const hoverIfNotDisabled = css<Pick<Props, 'disabled'>>`
-  ${p =>
-    p.disabled
-      ? ''
-      : `
-      &:hover:not(:focus-within) {
-        border-color: ${p.theme.color.inputBorderHover};
-      }
-`}
-`;
+const BottomWrapper = styled(motion.div)``;
 
-const focusBorderColor = css<Pick<Props, 'error'>>`
-  &:focus-within {
-    border-color: ${p => p.theme.color.borderActive};
+export const FormField: React.FC<Props> = ({
+  children,
+  className,
+  error,
+  extraInfo,
+  fieldId,
+  forId,
+  group,
+  hideLabel,
+  label,
+  required = false,
+  width,
+}) => {
+  const labelText = `${label} ${required ? '*' : ''}`;
+
+  const Field = () => {
+    if (group) {
+      return (
+        <Fieldset>
+          <Legend styleType="label">{labelText}</Legend>
+          {children}
+        </Fieldset>
+      );
+    }
+
+    if (fieldId || forId) {
+      return (
+        <>
+          <FormLabel hideLabel={hideLabel} forId={fieldId || forId}>
+            {labelText}
+          </FormLabel>
+          {children}
+        </>
+      );
+    }
+
+    return (
+      <>
+        <FormLabel hideLabel={hideLabel}>
+          {labelText}
+          {children}
+        </FormLabel>
+      </>
+    );
+  };
+
+  if (forId) {
+    assert(false, `FormField: the prop forId is deprecated, please use fieldId instead.`, {
+      level: 'warn',
+    });
   }
-  &.focus-within {
-    border-color: ${p => p.theme.color.borderActive};
-  }
-`;
 
-const outerFlexboxBorderColor = css<Pick<Props, 'error' | 'success'>>`
-  border: 1px solid
-    ${p => {
-      if (hasError(p.error)) return p.theme.color.inputBorderError;
-      if (p.success) return p.theme.color.inputBorderSuccess;
-      return p.theme.color.inputBorder;
-    }};
-  ${focusBorderColor}
-  ${hoverIfNotDisabled}
-`;
-
-const DensedTypography = styled(Typography)`
-  line-height: ${LINE_HEIGHT_INFO_BELOW}px;
-  display: inline-block;
-`;
-
-const InlineFlexbox = styled(Flexbox)`
-  display: flex;
-  width: 100%;
-`;
-
-const BottomWrapper = styled(motion.span)`
-  width: 100%;
-`;
-
-const CleanFlexbox = React.forwardRef((props: any, ref: React.Ref<HTMLDivElement>) => (
-  <Flexbox {...R.omit(['error', 'disabled', 'innerWrapperRef'], props)} ref={ref} />
-));
-
-const InnerWrapperFlexbox = styled(CleanFlexbox)`
-  position: relative;
-  ${height}
-  box-sizing: border-box;
-
-  ${outerFlexboxBorderColor}
-
-  background-color: #ffffff;
-  box-shadow: 0 1px 3px ${p => p.theme.color.shadowInput};
-  font: inherit;
-  font-size: inherit;
-  line-height: inherit;
-  color: inherit;
-`;
-
-const HidableTypography = styled(Typography)<{ hidden: boolean }>`
-  ${p => (p.hidden ? visuallyHidden : '')}
-  width: 100%;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-`;
-
-export const FormField: React.FC<Props> = React.forwardRef<HTMLDivElement, Props>(
-  (
-    {
-      children,
-      className,
-      disabled,
-      error,
-      extraInfo,
-      hideLabel,
-      innerWrapperRef,
-      label,
-      onClick,
-      required,
-      size,
-      success,
-      width,
-      fullWidth,
-    },
-    ref,
-  ) => (
-    /* eslint-disable jsx-a11y/label-has-associated-control,jsx-a11y/label-has-for */
-    <Wrapper width={fullWidth ? '100%' : width} className={className} onClick={onClick} ref={ref}>
-      <label>
-        <InlineFlexbox container direction="column">
-          <HidableTypography
-            hidden={Boolean(hideLabel)}
-            type="secondary"
-            color={t => t.color.label}
-          >
-            {label}
-            {required ? ' *' : ''}
-          </HidableTypography>
-          <Flexbox item alignSelf="stretch">
-            <Typography type="secondary" color={t => t.color.text} as="div">
-              <InnerWrapperFlexbox
-                ref={innerWrapperRef}
-                container
-                alignItems="center"
-                {...{ size, error, success, disabled }}
-              >
-                {children}
-              </InnerWrapperFlexbox>
-            </Typography>
-          </Flexbox>
-        </InlineFlexbox>
-      </label>
+  return (
+    <Wrapper width={width} className={className}>
+      <Field />
       <AnimatePresence>
         {hasError(error) ? (
           <BottomWrapper
@@ -146,30 +77,26 @@ export const FormField: React.FC<Props> = React.forwardRef<HTMLDivElement, Props
             exit={{ y: 0, opacity: 0 }}
             aria-live="polite"
           >
-            <DensedTypography type="tertiary" color={t => t.color.negative}>
+            <Typography type="tertiary" color={t => t.color.negative}>
               <VisuallyHidden>Error: </VisuallyHidden>
               {error}
-            </DensedTypography>
+            </Typography>
           </BottomWrapper>
         ) : (
           extraInfo && (
             <BottomWrapper
-              size={size}
               initial={{ y: -10, opacity: 0 }}
               animate={{ y: 0, opacity: 1 }}
               exit={{ y: 0, opacity: 0 }}
               aria-live="polite"
-              // TODO: Check this one @manman
-              // @ts-ignore
-              aria-relevant="additions removals"
             >
-              <DensedTypography type="tertiary" color={t => t.color.label}>
+              <Typography type="tertiary" color={t => t.color.label}>
                 {extraInfo}
-              </DensedTypography>
+              </Typography>
             </BottomWrapper>
           )
         )}
       </AnimatePresence>
     </Wrapper>
-  ),
-);
+  );
+};