+++ src/Molecules/Input/Text/Text.stories.tsx
@@ -51,6 +51,9 @@ storiesOf('Molecules | Input / Text', module)
   ))
   .add('With auto focus', () => <Input.Text label="Label" placeholder="Placeholder" autoFocus />)
   .add('Required', () => <Input.Text label="Label" placeholder="Placeholder" required />)
+  .add('Required with star', () => (
+    <Input.Text label="Label" placeholder="Placeholder" required visuallyEmphasiseRequired />
+  ))
   .add('Actions', () => (
     <>
       <p>