+++ src/Molecules/Input/Text/Text.types.ts
@@ -28,6 +28,7 @@ export type Props = {
   value?: string;
   defaultValue?: string;
   required?: boolean;
+  visuallyEmphasiseRequired?: boolean;
 
   onChange?: React.ChangeEventHandler<HTMLInputElement>;
   onClick?: React.MouseEventHandler<HTMLInputElement>;